[backup] trivial: extract command_adapter::Command to separate file
